-Corrections to add,delete and stats -patents are now strings -add,delete now work on real time -added web as a posible publish -Instagram is now only for Stories -Default Image for card -Publish Story button -Alerts for delete and add process

+++ src/app/dashboard/car-card/car-card.component.ts
@@ -16,10 +16,20 @@ export class CarCardComponent {
   carService: CarsService = inject(CarsService);
 
   DeleteCar() {
-    this.carService.deleteCar(this.car.patent);
+    if (confirm("Esta seguro que desea eliminar este auto?")) {
+      this.carService.deleteCar(this.car.patent);
+      alert("Auto eliminado con exito");
+    }
   }
 
   EditCar() {
     this.editCar.emit(this.car);
+  }
+
+  SubirHistoriaCar() {
+    if (confirm("Esta seguro que desea publicar una historia en Instagram de este auto?")) {
+      //this.carService.subirHistoriaInstagram(this.car.patent);
+      alert("Historia publicada con exito");
+    }
   }
 }
